+++ src/dataflow/operators/inspect.rs
@@ -92,3 +92,46 @@ where
         self.as_ref().map(|stream| stream.debug_inspect(inspect))
     }
 }
+
+// TODO: InspectExt for arrangements
+
+pub trait DebugTuple {
+    fn debug(&self) -> Self;
+}
+
+macro_rules! impl_tuple {
+    ($(($($ty:ident,)+)),* $(,)?) => {
+        $(
+            #[allow(non_snake_case)]
+            impl<$($ty,)+> DebugTuple for ($($ty,)+)
+            where
+                $($ty: InspectExt,)+
+                $(<$ty as InspectExt>::Value: Debug,)+
+            {
+                fn debug(&self) -> Self {
+                    let ($($ty,)+) = self;
+                    ($($ty.debug(),)+)
+                }
+            }
+        )*
+    };
+}
+
+impl_tuple! {
+    (A,),
+    (A, B,),
+    (A, B, C,),
+    (A, B, C, D,),
+    (A, B, C, D, E,),
+    (A, B, C, D, E, F,),
+    (A, B, C, D, E, F, G,),
+    (A, B, C, D, E, F, G, H,),
+    (A, B, C, D, E, F, G, H, I,),
+    (A, B, C, D, E, F, G, H, I, J,),
+    (A, B, C, D, E, F, G, H, I, J, K,),
+    (A, B, C, D, E, F, G, H, I, J, K, L,),
+    (A, B, C, D, E, F, G, H, I, J, K, L, M,),
+    (A, B, C, D, E, F, G, H, I, J, K, L, M, N,),
+    (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O,),
+    (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P,),
+}